add ToolBarModule, NavigationBarModule, ViewingBarModule, InfoBarModule, SearchingBarModule; remove TabModule
