+++ FileManager/FileManager.pro.user
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!DOCTYPE QtCreatorProject>
-<!-- Written by QtCreator 13.0.0, 2024-08-21T22:14:42. -->
+<!-- Written by QtCreator 13.0.0, 2024-08-22T23:29:30. -->
 <qtcreator>
  <data>
   <variable>EnvironmentId</variable>
